refactor: Meter를 삭제하고 Distance라는 객체로 변경
